refactor: make `LinkLabelWithSourceAndDestination` re-usable
